Add adjustable timing & use long in Timer
Added TimingVars struct and added the option to set:
* Hit offset, to offset gameplay from the music in case of delay
* The hit window, good and OK values
* The "dead window", so drums cannot be hit successfully by just
  spamming keys
Also, changed to store the raw millisecond value in Timer with long,
to facilitate some future changes in how Timer works and handle
longer periods of time. SDL's value is unsigned 32-bit, but we also
need to handle negative values, thus 32-bit signed int is not enough
for when the game is run for longer periods of time, like in an
arcade machine.