Add timing variable tuning to menus
Added menu entries with ValueDials for tuning the timing values in
timings.json.
Fixed missing initialisation of currentValue in ValueDial.
Added parameter to the callback in ValueDial (which it calls with
currentValue every time.)
Added localisation for the different menu entries.
Improved error handling for locales that failed to load.