Remove static timers array from Timer, update doc
On a side note, this was motivated by the fact that builds generated
by gdc would crash on runtime due to a range error, because of working
around this unfeature and invoking undefined behavior.

On that note, gdc 9.2.1 has been tested and confirmed working for
compiling OpenTaiko as of this commit.