Fix Cosmetic not getting positioned properly
Added a missing entry to multiply the rect.x coordinate by scroll.
It now behaves properly in regards to its scroll speed.

Also changed how position is set for Bashable - instead of having it
set the same as the graphical position, it is now set to the actual
value, reducing the complexity of the implementation.