Fix !zoom handling and add !reset
Fixed the inability to set the zoom value without first declaring a
new offset. Changing the zoom value will fill in any missing beats
in the stanza with blanks before declaring a new offset, resetting the
index and finally changing the zoom value.

Also added !reset in order to prevent blanks being filled in, should
it be desired.